reorganizing app folder
