+++ tonepoem-app/.gitignore
@@ -19,5 +19,3 @@
 .env.production.local
 
 npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
