+++ tonepoem-app/package.json
@@ -4,9 +4,12 @@
   "private": true,
   "proxy": "http://localhost:8080",
   "dependencies": {
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
     "@fortawesome/fontawesome-svg-core": "^6.1.1",
     "@fortawesome/free-solid-svg-icons": "^6.1.1",
     "@fortawesome/react-fontawesome": "^0.1.18",
+    "@mui/material": "^5.7.0",
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^7.2.1",
@@ -21,7 +24,7 @@
     "react-google-login": "^5.2.2",
     "react-paginate": "^8.1.3",
     "react-player": "^2.10.0",
-    "react-router-dom": "^5.3.0",
+    "react-router-dom": "^5.3.1",
     "react-scripts": "^5.0.0",
     "react-wordcloud": "^1.2.7",
     "recharts": "^2.1.9",
