--- tonepoem-app/src/components/test.js
@@ -1,24 +1,0 @@
-import React, { Component } from "react";
-
-//Test Implemented by: Seyun Kim
-class App extends Component {
-    state = {
-        audio: new Audio(sound),
-        isPlaying: false,
-    };
-
-    playPause = () => {
-        let isPlaying = this.state.isPlaying;
-
-        if (isPlaying) {
-            this.state.audio.pause();
-        }
-        else{
-            this.state.audio.play();
-        }
-
-        this.setState({isPlaying: !isPlaying});
-    };
-}
-
-export default test;
